+++ CHANGELOG.rst
@@ -1,8 +1,25 @@
 =====================================================
-manuelraa.ansible_inventory_yaml_folder Release Notes
+manuelraa.yaml_folder_inventory Release Notes
 =====================================================
 
 .. contents:: Topics
+
+
+v1.0.1
+======
+
+Breaking Changes
+----------------
+- Rename inventory plugin from :code:`manuelraa.inventory.yaml_folder` to :code:`manuelraa.yaml_folder.yaml_folder`
+
+Minor Changes
+-------------
+- Add type hints and basic docstrings
+
+Bugfixes
+--------
+- Removed debug print from plugin
+- Fix empty host vars in :code:`main.yml` causes error because parsed as None
 
 
 v1.0.0